tests/wpt/tests/html/canvas/element/canvas-host/2d.canvas.host.size.attributes.reflect.setidlzero.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.canvas.host.size.attributes.reflect.setidlzero</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.canvas.host.size.attributes.reflect.setidlzero</h1>
<p class="desc">Setting IDL attributes to 0 updates IDL and content attributes</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Setting IDL attributes to 0 updates IDL and content attributes");
_addTest(function(canvas, ctx) {

  canvas.width = 0;
  canvas.height = 0;
  _assertSame(canvas.getAttribute('width'), '0', "canvas.getAttribute('width')", "'0'");
  _assertSame(canvas.getAttribute('height'), '0', "canvas.getAttribute('height')", "'0'");
  _assertSame(canvas.width, 0, "canvas.width", "0");
  _assertSame(canvas.height, 0, "canvas.height", "0");

});
</script>
